Server Backend copy/requirements.txt
Django>=3.2
gunicorn
pymongo
requests
google-generativeai
schedule
feedparser
